requirements.txt
five==0.4.0
django-geonode-client==0.0.22
django-cors-headers==1.3.1
git+git://github.com/ftf-geonode/geonode.git@ftf
pymemcache
defusedxml
django-floppyforms
